+++ woocommerce-gateway-billmate/class-billmate-invoice.php
@@ -844,7 +844,7 @@ parse_str($_POST['post_data'], $datatemp);
 				'artnr'    => $sku,
 				'title'    => $item['name'],
 				'aprice'    =>  ($discount) ? ($billmate_item_standard_price*100) : ($priceExcl*100), //+$item->unittax
-				'taxrate'      => (float)$item_tax_percentage,
+				'taxrate'      => (int)$item_tax_percentage,
 				'discount' => ($discount) ? round((1 - ($billmate_item_price_including_tax/$billmate_item_standard_price)) * 100) : 0,
 				'withouttax' => $item['qty'] * ($priceExcl*100)
 			);
@@ -877,7 +877,7 @@ parse_str($_POST['post_data'], $datatemp);
 					'artnr'    => "",
 					'title'    => sprintf(__('Discount %s%% tax', 'billmate'),round($key,0)),
 					'aprice'    => -($discountAmount*100), //+$item->unittax
-					'taxrate'      => $key,
+					'taxrate'      => (int)$key,
 					'discount' => (float)0,
 					'withouttax' => -($discountAmount*100)
 
@@ -902,7 +902,7 @@ parse_str($_POST['post_data'], $datatemp);
 
 			$orderValues['Cart']['Shipping'] = array(
 				'withouttax'    => ($shipping_price-$order->order_shipping_tax)*100,
-				'taxrate'      => (float)$calculated_shipping_tax_percentage,
+				'taxrate'      => (int)$calculated_shipping_tax_percentage,
 
 			);
 			$total += ($shipping_price-$order->order_shipping_tax) * 100;
@@ -931,7 +931,7 @@ parse_str($_POST['post_data'], $datatemp);
 
 				$orderValues['Cart']['Handling'] = array(
 					'withouttax'    => $this->invoice_fee*100,
-					'taxrate'      => (float)$rate,
+					'taxrate'      => (int)$rate,
 				);
 
 				$total += $this->invoice_fee * 100;
@@ -986,7 +986,7 @@ parse_str($_POST['post_data'], $datatemp);
 				$rate = $rate['rate'];
 				$orderValues['Cart']['Handling'] = array(
 					'withouttax'    => $this->invoice_fee*100,
-					'taxrate'      => (float)$rate,
+					'taxrate'      => (int)$rate,
 				);
 
 				$total += $this->invoice_fee * 100;
